refactor(identity.jwks): TokenProvider 위치 이동
- 토큰 생성 방법에 대한 인터페이스 정의는 실제 구현체에서 제공하도록 변경
