+++ identity/identity-application/src/main/java/com/seatliberator/seatliberator/identity/infrastructure/security/authentication/method/token/TokenConfiguration.java
@@ -1,6 +1,6 @@
 package com.seatliberator.seatliberator.identity.infrastructure.security.authentication.method.token;
 
-import com.seatliberator.seatliberator.identity.core.token.TokenProvider;
+import com.seatliberator.seatliberator.jwks.application.port.in.TokenProvider;
 import com.seatliberator.seatliberator.identity.infrastructure.security.authentication.method.token.handler.DefaultTokenIssueProcessor;
 import com.seatliberator.seatliberator.identity.infrastructure.security.authentication.method.token.handler.TokenIssueProcessor;
 import org.springframework.context.annotation.Bean;
