+++ identity/identity-application/src/main/java/com/seatliberator/seatliberator/jwks/application/service/OpaqueTokenProvider.java
@@ -1,6 +1,6 @@
 package com.seatliberator.seatliberator.jwks.application.service;
 
-import com.seatliberator.seatliberator.identity.core.token.TokenProvider;
+import com.seatliberator.seatliberator.jwks.application.port.in.TokenProvider;
 import com.seatliberator.seatliberator.jwks.application.port.out.ByteEncoder;
 import lombok.RequiredArgsConstructor;
